% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphlan.R
\name{graphlan}
\alias{graphlan}
\title{Visualizing the results of \code{bouth} with GraPhlAn}
\usage{
graphlan(bouth.out, show.leaf = FALSE, output.dir = getwd(),
  graphlan.dir = NULL, clade.separation = NULL, branch.thickness = NULL,
  branch.bracket.depth = NULL, branch.bracket.width = NULL,
  clade.marker.size = NULL)
}
\arguments{
\item{bouth.out}{an output object from the \code{bouth} function.}

\item{show.leaf}{a logical value indicating whether or not the leaf nodes are
shown in the output figure.}

\item{output.dir}{the directory for storing the output figure. The default is
the current working directory.}

\item{graphlan.dir}{the directory where the GraPhlAn package is located.}
}
\description{
Create a tree figure for visualizing the results of \code{bouth}
}
\note{
This function is dependent on the python package GraPhlAn. Parameters
  \code{clade.separation}, \code{branch.thickness},
  \code{branch.bracket.depth}, \code{branch.bracket.width}, and
  \code{clade.marker.size} can be set the same as in the python package. Details can be found
  at
  (\href{https://bitbucket.org/nsegata/graphlan/overview}{https://bitbucket.org/nsegata/graphlan/overview}).
}
\examples{
data(IBD)

## performing the one-stage, weighted bottom-up test on the IBD data
test.1 = bouth(anno.table = IBD$tax.table, pvalue.leaves = IBD$pvalue.otus,
na.symbol = "unknown", far = 0.1, is.weighted = TRUE)


## suppose the GraPhlAn package is located at graphlan_directory/
graphlan(bouth.out = test.1, graph.dir = graphlan_directory)

}
\references{
Asnicar, F, et al. "Compact graphical representation of
  phylogenetic data and metadata with GraPhlAn." PeerJ 3 (2015): e1029.
}
